Toggle RX/TX for MOD-RS485-ISO
An Olimex MOD-RS485-ISO Modbus interface board has no DE or RE pins, but
instead RX and/or TX can be enabled via the I2C bus. Add this command to
the dereSet() function.
The delay between TX enable and writing the Modbus request is increased
slightly. With the shorter delay the following RX enabled command would
often not work.